Moved getPayloads to async application step 1
