+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/App.kt
@@ -9,17 +9,12 @@ import arrow.core.raise.result
 import arrow.fx.coroutines.resourceScope
 import dev.reformator.stacktracedecoroutinator.runtime.DecoroutinatorRuntime
 import io.ktor.server.application.Application
-import io.ktor.server.auth.authenticate
 import io.ktor.server.netty.Netty
 import io.ktor.server.routing.routing
 import io.ktor.utils.io.CancellationException
 import io.micrometer.prometheusmetrics.PrometheusConfig
 import io.micrometer.prometheusmetrics.PrometheusMeterRegistry
 import kotlinx.coroutines.awaitCancellation
-import no.nav.emottak.ebms.persistence.Database
-import no.nav.emottak.ebms.persistence.ebmsDbConfig
-import no.nav.emottak.ebms.persistence.ebmsMigrationConfig
-import no.nav.emottak.ebms.persistence.repository.PayloadRepository
 import no.nav.emottak.ebms.processing.ProcessingService
 import no.nav.emottak.ebms.sendin.SendInService
 import no.nav.emottak.ebms.validation.DokumentValidator
@@ -35,10 +30,6 @@ fun main() = SuspendApp {
         DecoroutinatorRuntime.load()
     }
 
-    val database = Database(ebmsDbConfig.value)
-    database.migrate(ebmsMigrationConfig.value)
-
-    val payloadRepository = PayloadRepository(database)
     val processingClient = PayloadProcessingClient(scopedAuthHttpClient(EBMS_PAYLOAD_SCOPE))
     val processingService = ProcessingService(processingClient)
 
@@ -57,8 +48,7 @@ fun main() = SuspendApp {
                     ebmsProviderModule(
                         dokumentValidator,
                         processingService,
-                        sendInService,
-                        payloadRepository
+                        sendInService
                     )
                 }
             ).also { it.engineConfig.maxChunkSize = 100000 }
@@ -77,8 +67,7 @@ fun main() = SuspendApp {
 fun Application.ebmsProviderModule(
     validator: DokumentValidator,
     processing: ProcessingService,
-    sendInService: SendInService,
-    payloadRepository: PayloadRepository
+    sendInService: SendInService
 ) {
     val appMicrometerRegistry = PrometheusMeterRegistry(PrometheusConfig.DEFAULT)
 
@@ -94,9 +83,5 @@ fun Application.ebmsProviderModule(
         registerNavCheckStatus()
 
         postEbmsSync(validator, processing, sendInService)
-
-        authenticate(AZURE_AD_AUTH) {
-            getPayloads(payloadRepository)
-        }
     }
 }
