+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/Routes.kt
@@ -5,11 +5,9 @@ import io.ktor.http.HttpStatusCode
 import io.ktor.server.request.header
 import io.ktor.server.response.respond
 import io.ktor.server.routing.Route
-import io.ktor.server.routing.get
 import io.ktor.server.routing.post
 import no.nav.emottak.constants.SMTPHeaders
 import no.nav.emottak.ebms.model.signer
-import no.nav.emottak.ebms.persistence.repository.PayloadRepository
 import no.nav.emottak.ebms.processing.ProcessingService
 import no.nav.emottak.ebms.sendin.SendInService
 import no.nav.emottak.ebms.util.marker
@@ -28,8 +26,6 @@ import no.nav.emottak.util.marker
 import no.nav.emottak.util.retrieveLoggableHeaderPairs
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
-
-private const val REFERENCE_ID = "referenceId"
 
 @OptIn(ExperimentalUuidApi::class)
 fun Route.postEbmsSync(
@@ -134,53 +130,4 @@ fun Route.postEbmsSync(
             ex.parseAsSoapFault()
         )
     }
-}
-
-@OptIn(ExperimentalUuidApi::class)
-fun Route.getPayloads(
-    payloadRepository: PayloadRepository
-): Route = get("/api/payloads/{$REFERENCE_ID}") {
-    var referenceIdParameter: String? = null
-    val referenceId: Uuid?
-    // Validation
-    try {
-        referenceIdParameter = call.parameters[REFERENCE_ID]
-        referenceId = Uuid.parse(referenceIdParameter!!)
-    } catch (iae: IllegalArgumentException) {
-        logger().error("Invalid reference ID $referenceIdParameter has been sent", iae)
-        call.respond(
-            HttpStatusCode.BadRequest,
-            iae.getErrorMessage()
-        )
-        return@get
-    } catch (ex: Exception) {
-        logger().error("Exception occurred while validation of async payload request")
-        call.respond(
-            HttpStatusCode.BadRequest,
-            ex.getErrorMessage()
-        )
-        return@get
-    }
-
-    // Sending response
-    try {
-        val listOfPayloads = payloadRepository.getByReferenceId(referenceId)
-
-        if (listOfPayloads.isEmpty()) {
-            call.respond(HttpStatusCode.NotFound, "Payload not found for reference ID $referenceId")
-        } else {
-            call.respond(HttpStatusCode.OK, listOfPayloads)
-        }
-    } catch (ex: Exception) {
-        logger().error("Exception occurred while retrieving Payload: ${ex.localizedMessage} (${ex::class.qualifiedName})")
-        call.respond(
-            HttpStatusCode.InternalServerError,
-            ex.getErrorMessage()
-        )
-    }
-    return@get
-}
-
-fun Exception.getErrorMessage(): String {
-    return localizedMessage ?: cause?.message ?: javaClass.simpleName
 }
